Rename hydra heads (#903)
* refactor eqV2 heads

* refactor init_weights

* add output_name attribute to eqV2 heads

* fix deprecated registry names

* remove debug breakpoint

* add default name for rank2 head
